+++ packages/angular/src/components/input.component.ts
@@ -23,6 +23,7 @@ import '@synergy-design-system/components/components/input/input.js';
  * @since 2.0
  *
  * @dependency syn-icon
+ * @dependency syn-divider
  *
  * @slot label - The input's label. Alternatively, you can use the `label` attribute.
  * @slot prefix - Used to prepend a presentational icon or similar element to the input.
@@ -31,6 +32,8 @@ import '@synergy-design-system/components/components/input/input.js';
  * @slot show-password-icon - An icon to use in lieu of the default show password icon.
  * @slot hide-password-icon - An icon to use in lieu of the default hide password icon.
  * @slot help-text - Text that describes how to use the input. Alternatively, you can use the `help-text` attribute.
+ * @slot increment-number-stepper - An icon to use in lieu of the default increment number stepper icon.
+ * @slot decrement-number-stepper - An icon to use in lieu of the default decrement number stepper icon.
  *
  * @event syn-blur - Emitted when the control loses focus.
  * @event syn-change - Emitted when an alteration to the control's value is committed by the user.
@@ -49,6 +52,10 @@ import '@synergy-design-system/components/components/input/input.js';
  * @csspart clear-button - The clear button.
  * @csspart password-toggle-button - The password toggle button.
  * @csspart suffix - The container that wraps the suffix.
+ * @csspart stepper - The container that wraps the number stepper.
+ * @csspart decrement-number-stepper - The decrement number stepper button.
+ * @csspart increment-number-stepper - The increment number stepper button.
+ * @csspart divider - The divider between the increment and decrement number stepper buttons.
  */
 @Component({
   selector: 'syn-input',
@@ -232,7 +239,7 @@ to `text`.
   }
 
   /**
-* Hides the browser's built-in increment/decrement spin buttons for number inputs.
+* Hides the increment/decrement spin buttons for number inputs.
  */
   @Input()
   set noSpinButtons(v: SynInput['noSpinButtons']) {