+++ packages/docs/stories/components/input.stories.ts
@@ -207,17 +207,22 @@ export const Invalid: Story = {
   `,
 };
 
-/**
- * The type attribute controls the type of input the browser renders.
- */
 export const InputTypes: Story = {
+  parameters: {
+    docs: {
+      description: {
+        story: generateStoryDescription('input', 'types'),
+      },
+    },
+  },
   render: () => html`
-  <syn-input type="email" placeholder="Email"></syn-input><br/>
-  <syn-input type="number" placeholder="Number"></syn-input><br/>
-  <syn-input type="date" placeholder="Date"></syn-input>`,
+    <syn-input type="email" placeholder="Email"></syn-input><br/>
+    <syn-input type="number" placeholder="Number"></syn-input><br/>
+    <syn-input type="date" placeholder="Date"></syn-input>
+  `,
 };
 
-export const PrefixSuffixIcons: Story = {
+export const PrefixSuffixTextAndIcons: Story = {
   parameters: {
     docs: {
       description: {
@@ -227,58 +232,91 @@ export const PrefixSuffixIcons: Story = {
   },
   render: () => html`
   <syn-input placeholder="Small" size="small">
-    <syn-icon name="house" slot="prefix"></syn-icon>
-    <syn-icon name="chat" slot="suffix"></syn-icon>
+    <span slot="prefix">prefix</span>
+    <span slot="suffix">suffix</span>
   </syn-input>
   <br/>
   <syn-input placeholder="Medium" size="medium">
-    <syn-icon name="house" slot="prefix"></syn-icon>
-    <syn-icon name="chat" slot="suffix"></syn-icon>
+    <span slot="prefix">prefix</span>
+    <span slot="suffix">suffix</span>
   </syn-input>
   <br/>
   <syn-input placeholder="Large" size="large">
-    <syn-icon name="house" slot="prefix"></syn-icon>
-    <syn-icon name="chat" slot="suffix"></syn-icon>
+    <span slot="prefix">prefix</span>
+    <span slot="suffix">suffix</span>
+  </syn-input>
+  <br/>
+  <syn-input placeholder="Small" size="small">
+    <syn-icon name="wallpaper" slot="prefix"></syn-icon>
+    <syn-icon name="wallpaper" slot="suffix"></syn-icon>
+  </syn-input>
+  <br/>
+  <syn-input placeholder="Medium" size="medium">
+    <syn-icon name="wallpaper" slot="prefix"></syn-icon>
+    <syn-icon name="wallpaper" slot="suffix"></syn-icon>
+  </syn-input>
+  <br/>
+  <syn-input placeholder="Large" size="large">
+    <syn-icon name="wallpaper" slot="prefix"></syn-icon>
+    <syn-icon name="wallpaper" slot="suffix"></syn-icon>
   </syn-input>`,
 };
 
-/**
- * Use  to customize the way form controls are drawn.
- * This example uses CSS grid to position the label to the left of the control,
- * but the possible orientations are nearly endless.
- * The same technique works for inputs, textareas, radio groups, and similar form controls.
- */
 export const CustomizingLabelPosition: Story = {
+  parameters: {
+    docs: {
+      description: {
+        story: generateStoryDescription('input', 'label-position'),
+      },
+    },
+  },
   render: () => html`
-  <syn-input class="label-on-left" label="Name" help-text="Enter your name"></syn-input>
-  <syn-input class="label-on-left" label="Email" type="email" help-text="Enter your email"></syn-input>
-  <syn-textarea class="label-on-left" label="Bio" help-text="Tell us something about yourself"></syn-textarea>
+    <syn-input class="label-on-left" label="Name" help-text="Enter your name"></syn-input>
+    <syn-input class="label-on-left" label="Email" type="email" help-text="Enter your email"></syn-input>
+    <syn-textarea class="label-on-left" label="Bio" help-text="Tell us something about yourself"></syn-textarea>
 
-  <style>
-    .label-on-left {
-      --label-width: 3.75rem;
-      --gap-width: 1rem;
-    }
+    <style>
+      .label-on-left {
+        --label-width: 3.75rem;
+        --gap-width: 1rem;
+      }
 
-    .label-on-left + .label-on-left {
-      margin-top: var(--syn-spacing-medium);
-    }
+      .label-on-left + .label-on-left {
+        margin-top: var(--syn-spacing-medium);
+      }
 
-    .label-on-left::part(form-control) {
-      display: grid;
-      grid: auto / var(--label-width) 1fr;
-      gap: var(--syn-spacing-3x-small) var(--gap-width);
-      align-items: center;
-    }
+      .label-on-left::part(form-control) {
+        display: grid;
+        grid: auto / var(--label-width) 1fr;
+        gap: var(--syn-spacing-3x-small) var(--gap-width);
+        align-items: center;
+      }
 
-    .label-on-left::part(form-control-label) {
-      text-align: right;
-    }
+      .label-on-left::part(form-control-label) {
+        text-align: right;
+      }
 
-    .label-on-left::part(form-control-help-text) {
-      grid-column-start: 2;
-    }
-  </style>`,
+      .label-on-left::part(form-control-help-text) {
+        grid-column-start: 2;
+      }
+    </style>
+  `,
+};
+
+export const Stepper: Story = {
+  parameters: {
+    docs: {
+      description: {
+        story: generateStoryDescription('input', 'stepper'),
+      },
+    },
+  },
+  render: () => html`
+  <syn-input type="number" min="0" max="10" value="0"></syn-input>
+  <br/>
+  <syn-input type="number" min="0" max="10" value="2"></syn-input>
+  <br/>
+  <syn-input type="number" min="0" max="10" value="10"></syn-input>`,
 };
 
 // Bundled screenshot story
@@ -293,6 +331,7 @@ export const Screenshot: Story = generateScreenshotStory({
   Disabled,
   Sizes,
   InputTypes,
-  PrefixSuffixIcons,
+  PrefixSuffixTextAndIcons,
   CustomizingLabelPosition,
-}, 360);
+  Stepper,
+}, 500);